requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.44.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-crispy-forms==1.12.0
django-filter==2.4.0
idna==2.10
oauthlib==3.1.1
Pillow==8.3.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.6
